build(miniconda): make install script idempotent

+++ images/ubuntu/scripts/build/install-miniconda.sh
@@ -8,13 +8,17 @@
 # shellcheck disable=SC1091
 source "$HELPER_SCRIPTS"/etc-environment.sh
 
-# Install Miniconda
-curl -fsSL https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-"${ARCH}".sh -o miniconda.sh \
-    && chmod +x miniconda.sh \
-    && ./miniconda.sh -b -p /usr/share/miniconda \
-    && rm miniconda.sh
-
 CONDA=/usr/share/miniconda
+
+if [ ! -d "$CONDA" ]; then
+    curl -fsSL https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-"${ARCH}".sh -o miniconda.sh \
+        && chmod +x miniconda.sh \
+        && ./miniconda.sh -b -p "$CONDA" \
+        && rm miniconda.sh
+else
+    echo "Miniconda directory already exists at $CONDA. Skipping installation."
+fi
+
 set_etc_environment_variable "CONDA" "${CONDA}"
 
-ln -s $CONDA/bin/conda /usr/bin/conda
+ln -sf "$CONDA/bin/conda" /usr/bin/conda
